Create README.md
 Handling missing data is an important step of any data analysis. In real life, the study of given processes and phenomena inevitably encounters missing data, which should be subject to peprocessing analysis. To reveal the nature of the missing data in general, whether they are completely random, random or non-random. To think about what could be done with them so as not to distort the database drastically, but on the contrary to improve their suitability for objective analysis. This report provides an overview of possible tools for dealing with missing data. The emphasis is not on their research, which as a methodology was done in a previous report, but on working with them. Different methods of visualization, imputation and removal are shown in order to enrich the theory and practice of data refinement as a pre-processing step in the field of economic data and others.

The minimum wage in the EU and countries close to the EU is used as an example database. Since the minimum wage is formed either on the basis of collective labor contracts, state decisions or on the basis of the change in basic parameters such as the average wage, etc. , its use is for purely educational purposes.

In this paper, we will focus on the imputation of missing data and the applicability of the algorithms in the minimum wage example. The basic logarithms of Univariate and Multivariate Imputation will be used. Their behavior will be evaluated and visualized.